The four map levels
Best Buy clone:  the store front is on the right hand side

FloorMart_clone:  store front is on top

Bullseye_clone:  store front is on the left

Mall_clone:  store front is on the right

I think mall came out best, but that was the last one I did.  Also, I forgot to set it to XML, right now it's CSV.  Not sure why it won't let me change that as I was able to do so for the others.

These are somewhat crude.  I might need to find better tilesets unless you fellas think these do the trick, at least for now.  I wish I could get the alpha channel to work so it looks like the trees are on the asphalt instead of how they look right now

I modeled Best Buy from an actual map from Google Maps.  Looks better in that regard to the others.

FloorMart, thought of switching it up and making the store front up top.  Also we can make that the final level.